GIC: create GIC of arbitrary version
Up until now the logic behind the function for
creating a GIC device would try to create a V3 GIC
device  and if that failed, then a GICv2 device would get
created. For being able to test the GICV2 functionality, we
want to also be able to specify an arbitrary version.

Signed-off-by: Diana Popa <dpopa@amazon.com>