fdt: eliminate platform specific testing code
Inside the unit tests for the creation of the FDT
there was platform specific validation: the FDT resulting
on the spot would be checked against an expected FDT that
resulted on m6g platforms. This would result in the inability
to run the tests on a GICv2 system for example.

Signed-off-by: Diana Popa <dpopa@amazon.com>